refactor: use function instead of an object for create nedb
