+++ packages/insomnia/src/insomnia-data/node-src/database/database.test.ts
@@ -4,7 +4,7 @@ import type { BaseModel } from '../../../models';
 import * as models from '../../../models';
 import type { ChangeBufferEvent } from '../..';
 import { database as db } from '../..';
-import { _repairDatabase } from './database-nedb';
+import { repairDatabase } from './repair-database';
 
 describe('init()', () => {
   it('handles being initialized twice', async () => {
@@ -320,7 +320,7 @@ describe('_repairDatabase()', async () => {
     ]);
 
     // Run the fix algorithm
-    await _repairDatabase();
+    await repairDatabase();
 
     // Make sure things get adjusted
     const descendants2 = (await db.getWithDescendants(workspace)).map(d => ({
@@ -469,7 +469,7 @@ describe('_repairDatabase()', async () => {
       },
     ]);
     // Run the fix algorithm
-    await _repairDatabase();
+    await repairDatabase();
     // Make sure things get adjusted
     const descendants2 = (await db.getWithDescendants(workspace)).map(d => ({
       _id: d._id,
@@ -534,7 +534,7 @@ describe('_repairDatabase()', async () => {
     expect((await models.apiSpec.getByParentId(w2._id))?.fileName).toBe('New Document');
     expect((await models.apiSpec.getByParentId(w3._id))?.fileName).toBe('Unique name');
     // Run the fix algorithm
-    await _repairDatabase();
+    await repairDatabase();
     // Make sure things get adjusted
     expect((await models.apiSpec.getByParentId(w1._id))?.fileName).toBe('Workspace 1'); // Should fix
     expect((await models.apiSpec.getByParentId(w2._id))?.fileName).toBe('Workspace 2'); // Should fix
@@ -556,7 +556,7 @@ describe('_repairDatabase()', async () => {
     const newRepoWithoutSuffix = await models.gitRepository.create({
       uri: 'https://github.com/foo/bar',
     });
-    await _repairDatabase();
+    await repairDatabase();
     expect(await db.findOne(models.gitRepository.type, { _id: oldRepoWithSuffix._id })).toEqual(
       expect.objectContaining({
         uri: 'https://github.com/foo/bar.git',
