+++ packages/insomnia/src/insomnia-data/node-src/index.ts
@@ -1,1 +1,1 @@
-export { nedbDatabase, flushChangesImpl } from './database/database-nedb';
+export { createNedbDatabase, flushChangesImpl } from './database/database-nedb';
